Backend/requirements.txt
flask
flask-cors
pandas
openmeteo-requests
requests-cache
retry_requests
PyJWT
pymongo
